backend/requirements.txt
aiobotocore==2.13.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==1.0.0
anyio==4.4.0
anytree==2.12.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
autopage==0.5.2
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
botocore==1.34.131
bullet==2.2.0
cachetools==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.7.0
cmd2==2.4.3
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
dnspython==1.16.0
docutils==0.21.2
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2023.12.2
gcsfs==2023.12.2.post1
gnupg==2.3.1
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
imagesize==1.4.1
ipwhois==1.2.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.23
langsmith==0.0.87
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
openai==0.27.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pillow==10.4.0
pinecone-client==3.1.0
pinecone-datasets==0.7.0
pinecone-notebooks==0.1.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.2
prompt_toolkit==3.0.47
property-manager==3.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==11.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==1.10.17
pyflakes==3.2.0
Pygments==2.18.0
pymongo==3.11.0
pyperclip==1.9.0
pyproject-api==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-secrets==23.4.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
redis==5.0.8
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3fs==2023.12.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools-scm==8.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.32
stack-data==0.6.3
stevedore==5.2.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tox==4.17.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
verboselogs==1.7
virtualenv==20.26.3
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xkcdpass==1.19.9
yamlreader==3.0.4
yarl==1.9.4
